lambda/cv_processor/requirements.txt
# CV Processor Lambda dependencies
# Core extraction
python-docx>=0.8.11
pdfplumber>=0.10.3

# OCR (for triple OCR fallback - Tesseract requires system install)
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.0.0

# AWS SDK (provided by Lambda runtime, but pinned for local dev)
boto3>=1.34.0

# Cohere embed client
cohere>=5.0.0
